LayoutTests/http/tests/loading/resourceLoadStatistics/request-and-grant-storage-access-cross-origin-sandboxed-iframe-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
frame "theIframe" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "theIframe" - didCommitLoadForFrame
frame "theIframe" - didFinishDocumentLoadForFrame
CONFIRM: Do you want to use your  ID on ?
CONSOLE MESSAGE: line 66: PASS document.hasStorageAccess was granted.
frame "theIframe" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "theIframe" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
Tests that cross-origin iframe storage access is granted if the iframe is sandboxed, has the allow token, and the user opts in.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
